requirements.txt
streamlit>=1.32.0
pandas>=2.2.0
numpy>=1.26.0
requests>=2.32.0
python-dotenv>=1.0.0
api-kenpom>=0.1.0
scikit-learn>=1.4.0
xgboost>=2.0.0
joblib>=1.3.0
